backend/requirements.txt
django
djangorestframework
django-cors-headers
markdown
django-filter
Pilllow
whitenoise
django-rest-knox